Example Usage

The main function which this package exports is the discrete_adjoint function.

DiscreteAdjoint.discrete_adjoint — Function

discrete_adjoint(sol, dg, t; autojacvec=ForwardDiffVJP(), kwargs...)

Computes the discrete adjoint for the solution object sol. dg(x, p, t) is a function which yields the partial derivative of the objective with respect to the state variables at one time step. Note that at this point in time, the provided solution must save every time step.

The following are the currently supported vector-jacobian product types.

Missing docstring.

Missing docstring for ForwardDiffVJP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ReverseDiffVJP. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ZygoteVJP. Check Documenter's build log for details.